[EI-243] [web] refactor: Update import paths for indicator-related services and view models
