+++ web/__tests__/business/hooks/indicator-board-metadata/use-indicator-board-metadata-list-view-model.hook.test.tsx
@@ -8,7 +8,7 @@ import { resetAllStore } from '@/app/store/stores/reset-store';
 import { act } from 'react-dom/test-utils';
 import { SWRProviderWithoutCache } from '@/app/ui/components/util/swr-provider';
 import { resetMockDB } from '@/app/mocks/db';
-import { IndicatorBoardMetadata } from '@/app/business/services/view-model/indicator-board-metadata-view-model.service';
+import { IndicatorBoardMetadata } from '@/app/business/services/numerical-guidance/view-model/indicator-board-metadata-view-model.service';
 
 const wrapper = SWRProviderWithoutCache;
