+++ web/app/business/hooks/numerical-guidance/custom-forecast-indicator/use-custom-forecast-indicator-list-in-metadata.hook.ts
@@ -1,7 +1,7 @@
 import { useFetchCustomForecastIndicatorList } from '@/app/store/querys/numerical-guidance/custom-forecast-indicator.query';
 import { useMemo } from 'react';
 import { useIndicatorBoardMetadataViewModel } from '../indicator-board-metedata/use-indicator-board-metadata-view-model.hook';
-import { convertCustomForecastIndicatorsViewModel } from '../../../services/view-model/custom-forecast-indicator-view-model.service';
+import { convertCustomForecastIndicatorsViewModel } from '../../../services/numerical-guidance/view-model/custom-forecast-indicator-view-model.service';
 
 export const useCustomForecastIndicatorListInMetadata = (indicatorBoardMetadataId?: string) => {
   const { indicatorBoardMetadata } = useIndicatorBoardMetadataViewModel(indicatorBoardMetadataId);
