+++ web/app/business/hooks/numerical-guidance/custom-forecast-indicator/use-custom-forecast-indicator-list-view-model.hook.ts
@@ -4,7 +4,7 @@ import {
   useDeleteCustomForecastIndicator,
   useFetchCustomForecastIndicatorList,
 } from '@/app/store/querys/numerical-guidance/custom-forecast-indicator.query';
-import { convertCustomForecastIndicatorsViewModel } from '../../../services/view-model/custom-forecast-indicator-view-model.service';
+import { convertCustomForecastIndicatorsViewModel } from '../../../services/numerical-guidance/view-model/custom-forecast-indicator-view-model.service';
 import { useMemo } from 'react';
 import { usePending } from '@/app/ui/components/view/hooks/usePending.hook';
 import { IndicatorType } from '@/app/store/stores/numerical-guidance/indicator-list.store';
