+++ web/app/business/hooks/numerical-guidance/indicator-board-metedata/use-indicator-board-metadata-list-view-model.hook.ts
@@ -6,7 +6,7 @@ import {
 } from '../../../../store/querys/numerical-guidance/indicator-board-metadata.query';
 import { useFetchIndicatorBoardMetadataList } from '../../../../store/querys/numerical-guidance/indicator-board-metadata.query';
 import { useEffect, useMemo } from 'react';
-import { convertIndcatorBoardMetadataList } from '../../../services/view-model/indicator-board-metadata-view-model.service';
+import { convertIndcatorBoardMetadataList } from '../../../services/numerical-guidance/view-model/indicator-board-metadata-view-model.service';
 import { usePending } from '@/app/ui/components/view/hooks/usePending.hook';
 import {
   IndicatorInMetadataUnitTypes,
