+++ web/app/business/hooks/numerical-guidance/indicator/use-live-indicators-value-view-model.hook.ts
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 import { LiveIndicatorRequestParams } from '../../../../store/querys/numerical-guidance/indicator.query';
-import { convertLiveIndicatorsValueViewModel } from '../../../services/view-model/indicator-value/actual-indicators-value-view-model.service';
+import { convertLiveIndicatorsValueViewModel } from '../../../services/numerical-guidance/view-model/indicator-value/actual-indicators-value-view-model.service';
 import { useIndicatorBoardMetadataViewModel } from '../indicator-board-metedata/use-indicator-board-metadata-view-model.hook';
 import { useIndicatorBoard } from '../indicator-board/use-indicator-board.hook';
 import { useIndicatorBoardMetadataStore } from '@/app/store/stores/numerical-guidance/indicator-board-metadata.store';
