+++ web/app/business/services/numerical-guidance/view-model/custom-forecast-indicator-view-model.service.ts
