+++ web/app/business/services/numerical-guidance/view-model/indicator-board-metadata-view-model.service.ts
