+++ web/app/business/services/numerical-guidance/view-model/indicator-list/indicator-view-model.service.ts
