+++ web/app/business/services/numerical-guidance/view-model/indicator-list/indicators/bond-indicator.service.ts
