+++ web/app/business/services/numerical-guidance/view-model/indicator-list/indicators/cryptocurrency-indicator.service.ts
