+++ web/app/business/services/numerical-guidance/view-model/indicator-list/indicators/etf-indicator.service.ts
