+++ web/app/business/services/numerical-guidance/view-model/indicator-list/indicators/forex-pair-indicator.service.ts
