+++ web/app/business/services/numerical-guidance/view-model/indicator-list/indicators/fund-indicator.service.ts
