+++ web/app/business/services/numerical-guidance/view-model/indicator-list/indicators/index-indicator.service.ts
