+++ web/app/business/services/numerical-guidance/view-model/indicator-list/indicators/indicator.service.ts
