+++ web/app/business/services/numerical-guidance/view-model/indicator-list/indicators/stock-indicator.service.ts
