+++ web/app/business/services/numerical-guidance/view-model/indicator-value/custom-forecast-indicator-value-view-model.service.ts
