+++ web/app/ui/components/numerical-guidance/custom-forecast-indicator/custom-forecast-indicator-list/custom-forecast-indicator-list.tsx
@@ -2,7 +2,7 @@ import { useCustomForecastIndicatorListViewModel } from '@/app/business/hooks/nu
 import { useSelectedIndicatorBoardMetadata } from '@/app/business/hooks/numerical-guidance/indicator-board-metedata/use-selected-indicator-board-metadata-view-model.hook';
 import React from 'react';
 import List from '../../../view/molecule/list';
-import { CustomForecastIndicator } from '@/app/business/services/view-model/custom-forecast-indicator-view-model.service';
+import { CustomForecastIndicator } from '@/app/business/services/numerical-guidance/view-model/custom-forecast-indicator-view-model.service';
 import CustomForecastIndicatorListItem from './custom-forecast-indicator-list-item';
 import Pending from '../../../view/molecule/pending';
 import { cn } from '@/app/utils/style';
