+++ web/app/ui/components/numerical-guidance/indicator-board-metadata/metadata-list/metadata-list.tsx
@@ -6,7 +6,7 @@ import Button from '../../../view/atom/button/button';
 import Pending from '../../../view/molecule/pending';
 import { useSelectedIndicatorBoardMetadata } from '@/app/business/hooks/numerical-guidance/indicator-board-metedata/use-selected-indicator-board-metadata-view-model.hook';
 import MetadataListItem from './metadata-list-item';
-import { IndicatorBoardMetadata } from '@/app/business/services/view-model/indicator-board-metadata-view-model.service';
+import { IndicatorBoardMetadata } from '@/app/business/services/numerical-guidance/view-model/indicator-board-metadata-view-model.service';
 import { PlusIcon } from '@heroicons/react/solid';
 
 const MetadataList = React.memo(function MetadataList() {
