+++ web/app/ui/components/numerical-guidance/indicator/indicator-list/indicator-list-result.tsx
@@ -1,5 +1,5 @@
 import { useIndicatorListByType } from '@/app/business/hooks/numerical-guidance/indicator/use-indicator-list-by-type.hook';
-import { Indicator } from '@/app/business/services/view-model/indicator-list/indicators/indicator.service';
+import { Indicator } from '@/app/business/services/numerical-guidance/view-model/indicator-list/indicators/indicator.service';
 import { ListChildComponentProps } from 'react-window';
 import WindowList from '../../../view/molecule/window-list';
 import React from 'react';
